OutlinedTextField disabled colors function moved into utils package
