+++ app/src/main/java/com/nailorsh/repeton/features/newlesson/presentation/ui/NewLessonScreenFirst.kt
@@ -7,7 +7,10 @@ import androidx.compose.foundation.background
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.interaction.MutableInteractionSource
 import androidx.compose.foundation.layout.*
-import androidx.compose.material3.*
+import androidx.compose.material3.Button
+import androidx.compose.material3.ButtonDefaults
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
 import androidx.compose.runtime.*
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
@@ -70,6 +73,7 @@ fun NewLessonScreen(
             is NewLessonUiState.ErrorSaving -> {
                 subjectError = lessonState.uiState.error
             }
+
             NewLessonUiState.SuccessSaving -> onNavigateNext()
             else -> {}
         }
@@ -270,31 +274,6 @@ fun NewLessonScreen(
     }
 
 
-}
-
-@Composable
-fun getColorsForTextField(error: Boolean): TextFieldColors {
-    if (!error) {
-        return OutlinedTextFieldDefaults.colors(
-            disabledTextColor = MaterialTheme.colorScheme.onSurface,
-            disabledBorderColor = MaterialTheme.colorScheme.outline,
-            disabledPlaceholderColor = MaterialTheme.colorScheme.onSurfaceVariant,
-            disabledLabelColor = MaterialTheme.colorScheme.onSurfaceVariant,
-            disabledLeadingIconColor = MaterialTheme.colorScheme.onSurfaceVariant,
-            disabledTrailingIconColor = MaterialTheme.colorScheme.onSurfaceVariant,
-            disabledSupportingTextColor = MaterialTheme.colorScheme.onSurfaceVariant
-        )
-    } else {
-        return OutlinedTextFieldDefaults.colors(
-            disabledTextColor = MaterialTheme.colorScheme.error,
-            disabledBorderColor = MaterialTheme.colorScheme.error,
-            disabledPlaceholderColor = MaterialTheme.colorScheme.error,
-            disabledLabelColor = MaterialTheme.colorScheme.error,
-            disabledLeadingIconColor = MaterialTheme.colorScheme.error,
-            disabledTrailingIconColor = MaterialTheme.colorScheme.error,
-            disabledSupportingTextColor = MaterialTheme.colorScheme.error
-        )
-    }
 }
