+++ app/src/main/java/com/nailorsh/repeton/features/newlesson/presentation/ui/components/first/EndTimeTextField.kt
@@ -14,7 +14,7 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.input.ImeAction
 import androidx.compose.ui.text.input.KeyboardType
 import com.nailorsh.repeton.R
-import com.nailorsh.repeton.features.newlesson.presentation.ui.getColorsForTextField
+import com.nailorsh.repeton.core.util.getDisabledInteractiveOutlinedTextFieldColors
 import java.time.LocalTime
 import java.time.format.DateTimeFormatter.ofLocalizedTime
 import java.time.format.FormatStyle
@@ -36,7 +36,7 @@ fun EndTimeTextField(
                 indication = null
             ) { onClick() },
         enabled = false,
-        colors = getColorsForTextField(isError),
+        colors = getDisabledInteractiveOutlinedTextFieldColors(isError),
         readOnly = true,
         singleLine = true,
         value = time.format(ofLocalizedTime(FormatStyle.SHORT)),
@@ -55,6 +55,6 @@ fun EndTimeTextField(
             )
         },
 
-    )
+        )
 }
