+++ app/src/main/java/com/nailorsh/repeton/features/schedule/presentation/ui/components/Calendar.kt
@@ -8,7 +8,7 @@ import java.time.LocalDate
 import java.util.Calendar
 
 @Composable
-internal fun CalendarDatePicker(
+fun CalendarDatePicker(
     selectedDay: LocalDate,
     selectedDayUpdate: (LocalDate) -> Unit,
     showDatePickerUpdate: () -> Unit,
