+++ app/src/main/java/com/nailorsh/repeton/features/schedule/presentation/ui/components/DaySlider.kt
@@ -13,11 +13,12 @@ import androidx.compose.ui.res.dimensionResource
 import androidx.compose.ui.unit.dp
 import com.nailorsh.repeton.R
 import com.nailorsh.repeton.common.data.models.Lesson
+import com.nailorsh.repeton.features.schedule.presentation.ui.BASE_DATE
 import java.time.LocalDate
 
 @OptIn(ExperimentalFoundationApi::class)
 @Composable
-internal fun DaySlider(
+fun DaySlider(
     selectedDay: LocalDate,
     onDaySelected: (LocalDate) -> Unit,
     weekPagerState: PagerState,
