+++ app/src/main/java/com/nailorsh/repeton/features/schedule/presentation/ui/components/LessonCard.kt
@@ -21,7 +21,7 @@ import java.time.format.DateTimeFormatter
 import java.time.format.FormatStyle
 
 @Composable
-internal fun LessonCard(lesson: Lesson, onClick: (Lesson) -> Unit, modifier: Modifier = Modifier) {
+fun LessonCard(lesson: Lesson, onClick: (Lesson) -> Unit, modifier: Modifier = Modifier) {
 
     ElevatedCard(
         elevation = CardDefaults.cardElevation(
